# orp-frontend/apps/web-antd/public/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenrestyPlus 配置管理平台 API",
    "description": "面向多中心 OpenResty / Nginx 集群的配置与资源管理接口。\n\n**通用约定**\n- 响应统一包装：`{ code, message, data }`；`code=0` 表示成功，`code=-1` 表示失败（`message` 携带失败原因）\n- 鉴权：除登录外所有接口需携带 `Authorization: Bearer <token>`\n- 列表接口统一支持 `page` / `pageSize` 分页参数，返回 `PageResult` 结构\n- 时间格式：`YYYY-MM-DD HH:mm:ss`（如 `2026-09-28 10:20:00`）\n\n**业务规则**\n- 中心：存在任一关联资源（节点/HTTP/Stream/DNS/证书）时禁止删除\n- 节点：运行中（running）状态不可直接删除，需先暂停或停止；本地 Docker 节点状态操作会实际启动、暂停或停止容器\n- 发布：本地 Docker east-1/east-2/east-3 支持节点内语法校验、原子切换、reload、版本探针与回滚；生产 SSH 通道尚未配置\n- 指标：本地访问日志和节点探测为数据源；配置 OPENRESTY_GEOIP_DB_PATH 与中心经纬度后提供真实来源聚合\n- HTTP 监听：同中心同端口唯一；关联证书的适用中心范围必须包含该监听所属中心\n- Stream 服务：同中心同协议同端口唯一\n- 证书：列表接口不下发 privateKey，明文仅通过 detail 接口获取；被 HTTP 监听引用时禁止删除\n- 审计：所有写操作（新增/更新/删除/下线）自动记录审计日志",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "当前站点的 Go 控制面（由前端开发服务器代理）"
    }
  ],
...
  "paths": {
    "/orp/geoip-database": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "查询当前 GeoIP City 数据库状态",
        "operationId": "getGeoIPDatabase",
        "responses": {
          "200": {
            "description": "成功（code=0）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/GeoIPDatabaseInfo"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "失败（code=-1，401）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "example": {
                  "code": -1,
                  "message": "登录状态已失效",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/orp/geoip-database/import": {
      "post": {
        "tags": [
          "dashboard"
        ],
        "summary": "导入或替换 GeoIP City .mmdb 数据库",
        "operationId": "importGeoIPDatabase",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功（code=0）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/GeoIPDatabaseInfo"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "失败（code=-1，400）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "example": {
                  "code": -1,
                  "message": "文件无效、不是 City 数据库或超过大小限制",
                  "data": null
                }
              }
            }
          },
          "401": {
            "description": "失败（code=-1，401）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "example": {
                  "code": -1,
                  "message": "登录状态已失效",
                  "data": null
                }
              }
            }
          },
          "403": {
            "description": "失败（code=-1，403）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "example": {
                  "code": -1,
                  "message": "当前角色无写入权限",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/orp/rbac/users": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "分页查询用户（支持 keyword/roleId/status 筛选）",
        "operationId": "listRbacUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "用户名/姓名/邮箱/手机号"
          },
          {
            "name": "roleId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "enabled",
                "disabled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PageResult<RbacUserItem>"
          }
        }
      },
      "post": {
        "tags": [
          "rbac"
        ],
        "summary": "新增用户",
        "operationId": "createRbacUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "username/realName/roleIds/status/password/email/phone"
        },
        "responses": {
          "200": {
            "description": "RbacUserItem"
          }
        }
      }
    },
    "/orp/rbac/users/{id}": {
      "put": {
        "tags": [
          "rbac"
        ],
        "summary": "编辑用户（含角色绑定与状态）",
        "operationId": "updateRbacUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RbacUserItem"
          }
        }
      },
      "delete": {
        "tags": [
          "rbac"
        ],
        "summary": "删除用户（自身账号受保护）",
        "operationId": "deleteRbacUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/orp/rbac/users/{id}/reset-password": {
      "post": {
        "tags": [
          "rbac"
        ],
        "summary": "重置用户密码为初始密码",
        "operationId": "resetRbacUserPassword",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ resetTo, username }"
          }
        }
      }
    },
    "/orp/rbac/roles": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "分页查询角色（含 userCount）",
        "operationId": "listRbacRoles",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PageResult<RbacRoleItem>"
          }
        }
      },
      "post": {
        "tags": [
          "rbac"
        ],
        "summary": "新增角色",
        "operationId": "createRbacRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "code/name/description/status"
        },
        "responses": {
          "200": {
            "description": "RbacRoleItem"
          }
        }
      }
    },
    "/orp/rbac/roles/{id}": {
      "put": {
        "tags": [
          "rbac"
        ],
        "summary": "编辑角色（内置角色仅可改描述/名称）",
        "operationId": "updateRbacRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RbacRoleItem"
          }
        }
      },
      "delete": {
        "tags": [
          "rbac"
        ],
        "summary": "删除角色（内置/绑定用户时禁止）",
        "operationId": "deleteRbacRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/orp/rbac/roles/{id}/permissions": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "读取角色权限集合",
        "operationId": "getRbacRolePermissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ permissionKeys }"
          }
        }
      },
      "put": {
        "tags": [
          "rbac"
        ],
        "summary": "保存角色权限分配",
        "operationId": "assignRbacRolePermissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "permissionKeys: string[]"
        },
        "responses": {
          "200": {
            "description": "{ permissionKeys }"
          }
        }
      }
    },
    "/orp/publish": {
      "post": {
        "tags": [
          "publish"
        ],
        "summary": "发起配置下发（可选策略：全量/权重分批+金丝雀观察）",
        "operationId": "createPublish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "nodeIds: number[]；strategy: { mode: all|weighted, batchCount: 1-5, canary: { enabled, waitSeconds: 10|30|60|300 } }"
        },
        "responses": {
          "200": {
            "description": "{ batchId, total }"
          }
        }
      }
    },
    "/orp/publish/batches/{id}": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "查询下发批次实时进度（含 batchGroups/canary 倒计时）",
        "operationId": "getPublishBatch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PublishBatch"
          }
        }
      }
    },
    "/orp/publish/batches/{id}/abort": {
      "post": {
        "tags": [
          "publish"
        ],
        "summary": "中止下发（金丝雀观察期内取消后续批次）",
        "operationId": "abortPublishBatch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ id, phase, abortedCount }"
          }
        }
      }
    },
    "/orp/publish/batches/{id}/advance": {
      "post": {
        "tags": [
          "publish"
        ],
        "summary": "立即推进（跳过金丝雀观察等待期）",
        "operationId": "advancePublishBatch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ id, phase }"
          }
        }
      }
    },
...
    "/orp/nodes/{id}/status": {
      "put": {
        "tags": [
          "nodes"
        ],
        "summary": "修改节点运行状态",
        "description": "本地 Docker 节点实际执行容器启动 / 暂停 / 停止并核对状态；maintenance 暂停容器，running 恢复健康检查，操作后记录审计。",
        "operationId": "updateNodeStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "资源 ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "$ref": "#/components/schemas/NodeStatus"
                  }
                }
              },
              "example": {
                "status": "maintenance"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功（code=0）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Node"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "失败（code=-1，400）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "example": {
                  "code": -1,
                  "message": "目标状态不合法：仅支持 running/paused/stopped/maintenance",
                  "data": null
                }
              }
            }
          },
          "404": {
            "description": "失败（code=-1，404）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "example": {
                  "code": -1,
                  "message": "节点不存在或已被删除",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
    "/orp/nodes/{id}/agent/reload": {
      "post": {
        "tags": [
          "nodes"
        ],
        "summary": "向在线 Agent 节点派发固定 reload",
        "operationId": "queueNodeAgentReload",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "资源 ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "description": "仅超级管理员；只允许固定 reload 操作，任务有效期两分钟。",
        "responses": {
          "201": {
            "description": "已排队任务 ID"
          },
          "409": {
            "description": "失败（code=-1，409）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "example": {
                  "code": -1,
                  "message": "Agent 未登记、离线或已有未完成任务",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
    "/orp/nodes/{id}/latency-history": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "查询节点探活延迟历史",
        "operationId": "getNodeLatencyHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "资源 ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "延迟采样和统计摘要"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponseError": {
        "type": "object",
        "description": "统一失败响应（code=-1）",
        "properties": {
          "code": {
            "type": "integer",
            "example": -1
          },
          "message": {
            "type": "string",
            "example": "失败原因"
          },
          "data": {
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "NodeStatus": {
        "type": "string",
        "enum": [
          "running",
          "paused",
          "stopped",
          "maintenance"
        ],
        "description": "节点运行状态：运行中 / 已暂停 / 已停止 / 维护中"
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}